refactor: move rendering functions from audit command to report module
Move printJson, buildBanner, and printAuditFooter from command.ts to
report.ts, following the same display-module pattern used by
score/command.ts. Reduces audit command.ts imports from 11 to 8,
clearing the max-dependencies lint warning.
