+++ src/shell/commands/audit/report.ts
@@ -12,9 +12,11 @@ import type { FormatCtx } from "../../../analysis/audits/types.ts";
 import type { ListSection } from "../../ui/list.ts";
 import { colors } from "../../ui/log.ts";
 import { relative } from "node:path";
+import { renderFooter } from "../../ui/footer.ts";
 import { stripKeyPrefix } from "../../../utils/paths.ts";
 
 const NONE = 0;
+const JSON_INDENT = 2;
 
 const KIND_LABELS: Record<string, string> = {
   directory: "Directory",
@@ -111,4 +113,118 @@ function formatReport(
   });
 }
 
-export { formatReport };
+/**
+ * Serializes the audit report into a machine-readable JSON structure for programmatic consumers.
+ * @param root - Absolute project root for relative path display
+ * @param dbPath - Path to the snapshot database
+ * @param createdAt - Snapshot creation timestamp (ms since epoch) or null
+ * @param report - Ordered audit results with definitions and findings
+ * @param filterTerms - Category filter terms to apply before output
+ * @kuralCauses writes JSON to stdout
+ */
+function printJson(
+  root: string,
+  dbPath: string,
+  createdAt: number | null,
+  report: AuditReport,
+  filterTerms: string[],
+): void {
+  const filtered =
+    filterTerms.length > NONE
+      ? report.filter(({ definition }) =>
+          filterTerms.some((term) => definition.title.toLowerCase().includes(term)),
+        )
+      : report;
+
+  const audits = filtered.map(({ definition, findings }) => ({
+    name: definition.name,
+    title: definition.title,
+    count: findings.length,
+    findings,
+  }));
+
+  const total = audits.reduce((sum, a) => sum + a.count, NONE);
+  const output = {
+    snapshot: relative(root, dbPath) || dbPath,
+    createdAt: createdAt === null ? null : new Date(createdAt).toISOString(),
+    total,
+    audits,
+  };
+
+  console.log(JSON.stringify(output, null, JSON_INDENT));
+}
+
+/**
+ * Builds the audit banner metadata from run context.
+ * @param root - Absolute project root for relative path display
+ * @param dbPath - Path to the snapshot database
+ * @param createdAt - Snapshot creation timestamp (ms since epoch) or null
+ * @param total - Total number of identified issues
+ * @param filterTerms - Active category filter terms
+ * @param disabledAudits - Set of audit names that were skipped
+ * @returns Key-value record of banner fields for display
+ * @kuralPure
+ */
+function buildBanner(
+  root: string,
+  dbPath: string,
+  createdAt: number | null,
+  total: number,
+  filterTerms: string[],
+  disabledAudits: Set<string>,
+): Record<string, string> {
+  const takenOn = createdAt === null ? "unknown" : new Date(createdAt).toLocaleString();
+  const banner: Record<string, string> = {
+    "on snapshot": relative(root, dbPath) || dbPath,
+    "taken on": takenOn,
+  };
+  if (filterTerms.length > NONE) {
+    banner["filter by"] = filterTerms.join(", ");
+  }
+  if (disabledAudits.size > NONE) {
+    banner["disabled"] = [...disabledAudits].join(", ");
+  }
+  banner["identified issues"] = String(total);
+  return banner;
+}
+
+/**
+ * Closes the audit output with term definitions and suggested follow-up commands.
+ * @kuralPatterns commandFooter
+ * @kuralCauses writes footer sections to stdout
+ */
+function printAuditFooter(): void {
+  renderFooter(
+    [
+      { term: "Similarity", definition: "cosine similarity between embedding vectors (0–100%)" },
+      {
+        term: "Dominant / next",
+        definition: "highest and second-highest child-to-parent similarity",
+      },
+      {
+        term: "Identity-content",
+        definition: "alignment between a container's name and its actual contents",
+      },
+      { term: "Axis score", definition: "is-does linguistic measurement for descriptions" },
+      {
+        term: "clusters",
+        definition: "groups of semantically similar children suggesting a split",
+      },
+      { term: "fence", definition: "statistical threshold computed from group distribution" },
+      { term: "group", definition: "the peer set used as baseline for comparison" },
+    ],
+    [
+      { command: "kural audit -f <category>", description: "filter by audit category" },
+      { command: "kural audit -e", description: "show all findings per audit (no truncation)" },
+      { command: "kural audit -d <name>", description: "disable specific audits" },
+      { command: "kural audit -k <n>", description: "adjust sensitivity threshold" },
+      { command: "kural audit --json", description: "output result as JSON" },
+      {
+        command: "kural snapshot generate",
+        description: "regenerate snapshot after fixing issues",
+      },
+    ],
+  );
+}
+
+export { buildBanner, formatReport, printAuditFooter, printJson };
